logging-service/requirements.txt
fastapi~=0.73.0
uvicorn~=0.15.0
pydantic[dotenv]~=1.9.0
hazelcast-python-client==5.3.0